Move schema processing logic to conf.py
