+++ docs/reference/schema/hazard.md
@@ -9,7 +9,7 @@ depth: 1
 
 The `hazard` component is described as:
 
-```{jsoninclude-quote} ../../../schema/rdls_schema_processed.json
+```{jsoninclude-quote} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 ---
 jsonpointer: /properties/hazard/description
 ---
@@ -170,7 +170,7 @@ file: ../../../examples/hazard/aqueduct/Hazard metadata_Event sets_Hazards.csv
 
 ## Properties
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 ---
 pointer: /properties/hazard
 collapse: event_sets
@@ -197,7 +197,7 @@ externallinks: >-
         Event_set o|--|{ "Event": "Includes"
 ```
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 ---
 pointer: /properties/hazard/properties/event_sets/items
 collapse: hazards,events
@@ -225,7 +225,7 @@ externallinks: >-
         "Event_set" |o--o{ Event: "Includes"
 ```
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 ---
 pointer: /properties/hazard/properties/event_sets/items/properties/events/items
 collapse: hazard
@@ -238,7 +238,7 @@ externallinks: >-
 
 ## Hazard
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 ---
 pointer: /properties/hazard/properties/event_sets/items/properties/hazards/items
 ---
