+++ docs/reference/schema/index.md
@@ -18,7 +18,7 @@ The schema provides the authoritative definition of the structure of Risk Data L
 
 For this version of RDLS, the canonical URL of the schema is [https://docs.riskdatalibrary.org/en/0\_\_2\_\_0/rdls_schema.json](https://docs.riskdatalibrary.org/en/0__2__0/rdls_schema.json). Use the canonical URL to make sure that your software, documentation or other resources refer to the specific version of the schema with which they were tested.
 
-The pages in this section present the schema in tables with additional information in paragraphs. You can also [view the schema in an interactive browser](../browser.md) or [download it as JSON Schema](../../../schema/rdls_schema_processed.json).
+The pages in this section present the schema in tables with additional information in paragraphs. You can also [view the schema in an interactive browser](../browser.md) or [download it as JSON Schema](../../../docs/_readthedocs/html/rdls_schema_processed.json).
 
 ```{note}
    If any conflicts are found between the text in the documentation and the text within the schema, the text within the schema takes precedence.
