+++ docs/reference/schema/vulnerability.md
@@ -9,7 +9,7 @@ depth: 1
 
 The `vulnerability` component is described as:
 
-```{jsoninclude-quote} ../../../schema/rdls_schema_processed.json
+```{jsoninclude-quote} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 ---
 jsonpointer: /properties/vulnerability/description
 ---
@@ -119,7 +119,7 @@ file: ../../../examples/vulnerability/flood_jrc/Vulnerability metadata_Asset cos
 
 ## Properties
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability
 :collapse: functions/vulnerability,functions/fragility,functions/damage_to_loss,functions/engineering_demand,socio_economic
 :externallinks: >
@@ -134,7 +134,7 @@ file: ../../../examples/vulnerability/flood_jrc/Vulnerability metadata_Asset cos
 
 ## Vulnerability function
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability/properties/functions/properties/vulnerability/items
 :collapse: hazard_primary,hazard_secondary
 :externallinks: >
@@ -146,7 +146,7 @@ file: ../../../examples/vulnerability/flood_jrc/Vulnerability metadata_Asset cos
 
 ## Fragility function
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability/properties/functions/properties/fragility/items
 :collapse: hazard_primary,hazard_secondary
 :externallinks: >
@@ -158,7 +158,7 @@ file: ../../../examples/vulnerability/flood_jrc/Vulnerability metadata_Asset cos
 
 ## Damage-to-loss function
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability/properties/functions/properties/damage_to_loss/items
 :collapse: hazard_primary,hazard_secondary
 :externallinks: >
@@ -170,7 +170,7 @@ file: ../../../examples/vulnerability/flood_jrc/Vulnerability metadata_Asset cos
 
 ## Engineering demand function
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability/properties/functions/properties/engineering_demand/items
 :collapse: hazard_primary,hazard_secondary
 :externallinks: >
@@ -182,12 +182,12 @@ file: ../../../examples/vulnerability/flood_jrc/Vulnerability metadata_Asset cos
 
 ## Socio-economic indicator
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability/properties/socio_economic/items
 ```
 
 ## Hazard
 
-```{jsonschema} ../../../schema/rdls_schema_processed.json
+```{jsonschema} ../../../docs/_readthedocs/html/rdls_schema_processed.json
 :pointer: /properties/vulnerability/properties/functions/properties/vulnerability/items/properties/hazard_primary
 ```
